Stop pattern matching at the top level
Pattern matching was hanging with many instance methods. Upon inspection
of the AST, the only difference was an extra reference to the instance
method binding, with something like `$cfoo` for an instance method `foo`.
Likely, the pattern matching didn't play well with the new self-referential
frame bound to top-level modules. Note that adding an explicit extra
binding (like `foo = foo; bar = foo`) terminated just fine, so the problem
remains mysterious.

It wasn't root-caused because the symptom was easily, as we don't need to
pattern-match in the top-level anyways for anti-cycling or pattern matching
(the very rare cases that it actually does happen in the top-level). These
are both handled by frame updates in the one mandatory iteration of the
`reduce` iterant.