Remove AbsBinds aliasing from the top level
When reduction turns up empty due to recursion in the RHS of a top-level
binding, it is cut from the stack frame to be consistent with the truth
value of a failed pattern match. However, AbsBinds exports aliases to these
variables, and when they're absent, the AbsBind alias will resolve to
itself and recurse infinitely. This isn't caught by the HsLam visited check
because it's a pure substitution of that variable with itself.

This commit cuts out AbsBind aliasing from the top-level only -- the only
place it could cause trouble via soft binding. Despite being rather crude,
it is robust against changes to the lookup scheme, compared say to trying
to keep the symbol it points to alive in the stack.

PS: Worth noting the sheer luck of lookups succeeding up until this point.
Until now it seems every binding at the top-level has been populated (i.e.
only messing with recursion in the inner blocks) and that every AbsBind
alias is later in the Ord than the symbol it aliases. In light of the
recency of removing `pat_match` and instituting soft lookups at the top
level, the timing for this failure to finally crop up wasn't surprising,
but the mishmash of factors to keep it hidden sure was.