Correct Var and HsWrap handling
This commit involves two separate fixes:

1. Variables didn't always have their types substituted correctly/deeply
enough: usually the "don't recurse" behavior of `Var`'s `Data` instance was
getting in the way.
2. It was assumed that the type packaged with every HsExpr was the inferred
type of that callsite. It turns out instead that the packaged type is the
original type signature of the function/class binding that matched the
callsite. The callsite-specific types are stowed within an HsWrap (as well as
class/instance evidence stored in AbsBinds; see
<https://mail.haskell.org/pipermail/ghc-devs/2018-June/015936.html> for more
details). To lookup by type, it's necessary to make sure the same types share
the same name, and that the concrete specializations to the function types are
applied (e.g. `id 42` coerces `id :: forall a.a -> a` to `id :: Int -> Int` by
way of coercion (a ~ Int).

The HsWrapper solution/hack is very primitive for now, and may be extremely
expensive. It hangs unpredictably, which may mean that it's trying to deeply
substitute recursive types. The suspect that comes to mind is our very own
`blank_type`, which we may want to change to `Λ a :: TYPE`.

Additionally, this approach is quite redundant against GHC's functionality to
compile to Core, which applies these HsWrap. The migration to Core is just
too painful for now, but I don't rule it out as necessary for correctness
later.