Tie BindFrame in top-level
Top-level symbols with binding recursion are tamed by using the HsLam
anti-cycle logic, and looking for repeated callsites. This is because
all bindings that are relevant to the analysis are `FunBind`, even if
they have no arguments, and thus they're wrapped in `HsLam`. This has
a separate pathway to `HsVar` resolution, so the difference between a
wrapped and unwrapped symbol can lead to or prevent hangups.

Previously, if the top-level bind also has an explicit type signature
then the symbol would be resolved to itself with a Stack led by a
frame containing a bare HsVar reference, as a reduction on the top-
level binds. This stack frame led to the recursion on the case
presented in Z.hs, where in the bug case, the `deriving` clause
creates an instance of `foo :: a` referring to the binding of `foo`
for `F a`. This led to the infinite loop akin to the MVE in `Z.hs`.

However, when the top-level binding is knot-tied, the first frame is
always the unreduced top-level frame with the expressions wrapped in
lambdas.

PS. This suggests that the most sustainable strategy is to special-
case this frame as a separate record in `Stack`, but this will have
to wait until we establish a proper method to find the top-level binds
so that the initial SymApps can be assigned the Stack that is tied to
themselves.