Split TyConApps via splitAppTy
It turns out that TyConApps have at least two structural flavors:

1. `TyConApp TyCon [Type, ...]`
2. `TyAppTy (TyConApp TyCon []) (Ty...)`

It seems the former is used when the parameter[s?] is a TyVar, while
the latter is used when the parameter[s?] is a TyConApp itself. However,
it seems `splitAppTy` splits them identically, so it's better to leave all
the splitting to it.