Loosen `sub_sa_types_T` for empty arguments
Previously, empty arguments (caused by `error` and recursive terms)
caused the whole type substitution to fail because `inst_subty` was
used on all argument simultaneously (i.e. AND semantics). Now, we
treat each argument separately (OR semantics) because of we want to
handle `error` more gracefully. It's special-cased for separate
reasons because it introduces bare `HsLit` into the AST that
`get_expr_type` doesn't want to handle at the moment.