Preserve `forall` in `get_expr_type`
Foralls are now crucial parts of types because they are used for binding
with evidence/coercions via HsWrap. Therefore, it was necessary for
get_expr_type to preserve these because its result is used directly in
`unHsWrap`. Rather than stripping all foralls off the bat then, it
uses the new function `drop_ty_args` which unwraps and reconstructs the
forall clauses.