Add explicit handling for TyConApp in inst_subty
In 5ac54c56 I explained that I was assuming that AppTy would split
TyConApp in a consistent way that would eliminate the need to handle
TyCon. However, it seems that this is not true when there's a bare
TyCon without arguments, in which case splitTyConApp_maybe succeeds
but splitAppTy_maybe fails. It is therefore necessary to account for
the situation where the incoming type is a TyCon but fails the AppTy
split.