`reduce`: initially reduce_deep and match write sym locs
The reduce state machine was starting off on the wrong foot after 532e489f
when pattern matching (and hence initial reduction) at the top level was
removed. `expand_reads` then only saw zero-arg lambdas, and since it didn't
reduce_deep by itself, it didn't do anything either.

Additionally, the iterant was stopping too early because the locations
of pipes weren't being considered in the stop condition, only `sa_loc`.

PS I considered that maybe it was necessary to "progress" the state of syms in
the reduce iterant for correctness to properly handle chains of left-hand
dependencies on pipes, as seen in `R.hs` here. To avoid falling prey to the
same mistake, if one thinks that each iteration only unwraps one layer of
writes, which would then mandate a "progressive" iterant, it actually suffices
to only log statements because the reason there is a dependency chain is
because _writes_ from newly-resolved lambdas will saturate lambdas elsewhere.
This can be handled by subbing a growing bank of writes to the same initial
starter. On the other hand, "progressive" iteration is incorrect because when
a pipe dependency isn't resolved yet and has new values yet to come, its
reference to the pipe is already subbed for the values up to that point.

This partly explains why the process needs to be so slow. The n^4 write
uniqueness check isn't helping. Todo is to roll up sleeves and bring this down
to ideal (nlogn) with Map.