Make `is_monadic` a property of SymApps
The check to see if an expression was sent to the IO engine vs. remained
in pure land (i.e. "is_monadic") used to be implicit, checking if a
special `StmtFrame` stack frame came above all `AppFrame`s, which hinted
that the latest context for that expression was an "IO spine": a statement
in a `do` block or monadic operators for IO. This was combined with a
check in `pat_match` to see if it was the outermost statement (e.g.
arguments in `do` blocks aren't on the IO spine).

However, this test was severely flawed. In particular, the stack frame
over which `pat_match` evaluated that `is_first` condition didn't
necessarily have the `do` statement[s] as the outermost expression, e.g.
`f = runST $ do ...`. Furthermore, while StmtFrame seemed to be a
working necessary condition, it is sensitive to changes to the way
stack frames are handled, since it doesn't otherwise affect program
semantics. For example, the stack frame combination due to a
substitution would pose a threat to this logic.

On the other hand, whether or not an expression is monadic isn't a very
complex property: it's easy to flag (certain positions in certain
expressions) and easy to propagate (pass on to substituted value).
With so little and such straightforward maintenance, `is_monadic` as
a property for `SymApp` is sufficiently useful that it outweighs the
dirtiness it adds to the signature of SymApp and its increasing
state complexity.

PS this is vigorously a compromise in complexity, and the complexity of
SymApp is indeed a pressing issue. Rethinking its records and role is
justified further by the necessity of this state, so I expect that
experiments in SymApp form will start to appear as the datatype becomes
clumsier and clumsier to work with.