Progress in recursive read expansion for recursive pipe values
When writing a read directly back into the source pipe, the engine hangs.
This behavior arises often from `modify__`-type functions. This commit
unravels the expanded reads by 1 layer in the `reduce` `go` recursion.

This solution is not general, because if the read is wrapped in anything
(e.g. `var :: MVar [()]; readMVar var >>= putMVar . (():)`), then it
still hangs. Generally, either the read deduping needs to move into the
`go` recursion, or the recursion needs to be less zealous, e.g. expanding
reads after digging, or somehow only doing one layer at a time and moving
the progression to the iterant.