Stop expanding reads into stacks
This was causing a huge memory leak when encountering even slightly
complicated pipe writing situations. For example, the `scan` implementation
in `V.hs` has a single pipe write and two `do` blocks (one in the body of
`<*>`) and it leaks into the GB of memory.

This recursion was explicitly introduced in 88b8ed30, although without
enough supporting evidence. The tests pass even if frames aren't expanded
leading me to believe that perhaps the new progressive iterant might be
taking that burden.