Move read expansion upper bound to AST traversal
The AST traverser (i.e. `reduce`'s `go`) has the context to know exactly
how many actual expansions have been performed as it recurses into the
expansion results. The upper bound for this number of expansions is then
moved up to the traverser so it can recurse into expansion results and
avoid hanging.