Initial source plugin
This is the only viable way it seems to run ra.hs on other codebases.
The GHC API in ra.hs needs to know where to look for Cabal packages,
and as well, all the typechecked ASTs need to be coalesced to have
access to all the top-level variables.

1. The former has been hard because the only other solution has been
[ezyang's](http://blog.ezyang.com/2017/02/how-to-integrate-ghc-api-programs-with-cabal/)
using front-end plugins. However, re: <https://mail.haskell.org/pipermail/ghc-devs/2020-January/018433.html>
this is problematic so far on Cabal 3 & GHC 8.6.5. Source plugins
also have a nice workflow advantage, in that they slip right into the
build pipeline without needing to run compilation on the target
modules explicitly.
2. The latter has been hard because the GHC AST really doesn't want to
be serialized. It's not `Generic` so most of the modern serializers
won't work. None of the SYB-based serializers want to work either for
various reasons: GHC.Compact won't work because somewhere there's a
"pinned" type. Packman can't serialize the MVar that FastStrings hold
to the gloal hash table.
  - This is still a problem, although no longer for correctness. It's
    causing the typechecking of `purebase` to bottleneck the build
    process.