HsLam alt handling: and -> or over alts
Previously, all alt bindings were lumped and matched with `and` to combine
results. This is evidently wrong, since a) the bindings shouldn't be lumped
together because b) alts can match on their own and be evaluated while others
fail. Instead, each alt is matched separately, and evaluated separately based
on that match; the ReduceSyms are then combined.